refactor(legacy-mbti): split attempt service into pack/report/lifecycle services
